refactor: unify rule selection test
